Added icon to `www` folder for windows Renamed build-backend to build
