+++ packages/admin/test/mocha.setup.js
@@ -3,7 +3,7 @@ process.on('unhandledRejection', e => {
     throw e;
 });
 
-// enable the should interface with sinon
+// enable the "should" interface with sinon
 // and load chai-as-promised and sinon-chai by default
 const sinonChai = require('sinon-chai');
 const chaiAsPromised = require('chai-as-promised');
